py-srv/bin/requirements.txt
bottle
bottle-sqlalchemy
pymysql

